1.1.1 test
Iteration 1, user story 1, step 1.

It is the first step of the exercise.

Note that only one test method is implemented, the test method for the negative case is a stub.
This way, we don't write too many tests before starting the implementation.
The goal is to write a bit of test and a bit of code iteratively.